requirements.txt
contextlib2==0.5.5
importlib-metadata==0.5
path.py==11.3.0
pathlib2==2.3.2
PyYAML==3.13
scandir==1.9.0
six==1.11.0
